requirements.txt
pyyaml
numpy
neo4j
openai
langchain-openai
langchain
PyPDF2
python-docx
tenacity
docx
fastapi
pydantic
uvicorn
tiktoken
redis
celery
flower
pydantic-settings